fix(apple): allow each tab to have their own navigation stack, split profile page from profile tab view, remove dead route
