+++ apple/TasteNotes/UI/AppRouter.swift
@@ -67,7 +67,6 @@ enum Route: Hashable {
   case settings
   case currentUserFriends
   case friends(Profile)
-  case activity(Profile)
   case addProduct(Barcode?)
 }
 
@@ -83,21 +82,6 @@ public struct WithRoutes<RootView: View>: View {
 
   public var body: some View {
     view()
-      .navigationDestination(for: CheckIn.self) { checkIn in
-        CheckInScreenView(checkIn: checkIn)
-      }
-      .navigationDestination(for: Profile.self) { profile in
-        ProfileTabView(profile: profile)
-      }
-      .navigationDestination(for: Product.Joined.self) { product in
-        ProductScreenView(product: product)
-      }
-      .navigationDestination(for: Location.self) { location in
-        LocationScreenView(location: location)
-      }
-      .navigationDestination(for: Company.self) { company in
-        CompanyScreenView(company: company)
-      }
       .navigationDestination(for: Route.self) { route in
         switch route {
         case let .company(company):
@@ -106,8 +90,6 @@ public struct WithRoutes<RootView: View>: View {
           FriendsScreenView(profile: profileManager.getProfile())
         case .settings:
           PreferencesScreenView()
-        case let .activity(profile):
-          ActivityTabView(profile: profile)
         case let .location(location):
           LocationScreenView(location: location)
         case let .profileProducts(profile):
@@ -117,7 +99,7 @@ public struct WithRoutes<RootView: View>: View {
         case let .checkIn(checkIn):
           CheckInScreenView(checkIn: checkIn)
         case let .profile(profile):
-          ProfileTabView(profile: profile)
+          ProfileScreenView(profile: profile)
         case let .product(product):
           ProductScreenView(product: product)
         case let .friends(profile):
